Refactor #11: Add FoodListView file
- FoodListViewController에서 FoodListView 분리
- FoodListViewController에서 Model 분리
- FoodListViewController에서 NavigationUI관련 코드 extension 분리
